add logic for proof da inclusion
